+++ ui/src/app/core/accessories/types/humidifierdehumidifier/humidifierdehumidifier.manage.component.ts
@@ -27,7 +27,8 @@ export class HumidifierDehumidifierManageComponent implements OnInit {
   $activeModal = inject(NgbActiveModal)
 
   @Input() public service: ServiceTypeX
-  public targetMode: number | 'off'
+  public targetState: number
+  public targetMode: number
   public targetHumidityChanged: Subject<any> = new Subject<any>()
 
   public RelativeHumidityDehumidifierThreshold: CharacteristicType
@@ -39,35 +40,29 @@ export class HumidifierDehumidifierManageComponent implements OnInit {
 
   constructor() {
     this.targetHumidityChanged
-      .pipe(
-        debounceTime(300),
-      )
+      .pipe(debounceTime(300))
       .subscribe(() => {
-        switch (this.targetMode) {
-          case 0:
-            // auto
-            this.service.getCharacteristic('RelativeHumidityHumidifierThreshold').setValue(this.autoHumidity[0])
-            this.service.getCharacteristic('RelativeHumidityDehumidifierThreshold').setValue(this.autoHumidity[1])
-            break
-          case 1:
-            // Humidifier
-            this.service.getCharacteristic('RelativeHumidityHumidifierThreshold').setValue(this.targetHumidifierHumidity)
-            break
-          case 2:
-            // Dehumidifier
-            this.service.getCharacteristic('RelativeHumidityDehumidifierThreshold').setValue(this.targetDehumidifierHumidity)
-            break
+        if (this.RelativeHumidityHumidifierThreshold) {
+          this.service.getCharacteristic('RelativeHumidityHumidifierThreshold').setValue(this.targetHumidifierHumidity)
+        }
+        if (this.RelativeHumidityDehumidifierThreshold) {
+          this.service.getCharacteristic('RelativeHumidityDehumidifierThreshold').setValue(this.targetDehumidifierHumidity)
         }
       })
   }
 
   ngOnInit() {
-    this.targetMode = this.service.values.Active ? this.service.values.TargetHumidifierDehumidifierState : 'off'
-
+    this.targetState = this.service.values.Active
+    this.targetMode = this.service.values.TargetHumidifierDehumidifierState
     this.RelativeHumidityDehumidifierThreshold = this.service.getCharacteristic('RelativeHumidityDehumidifierThreshold')
     this.RelativeHumidityHumidifierThreshold = this.service.getCharacteristic('RelativeHumidityHumidifierThreshold')
-
     this.loadTargetHumidity()
+    setTimeout(() => {
+      const sliderElements = document.querySelectorAll('.noUi-target')
+      sliderElements.forEach((sliderElement: HTMLElement) => {
+        sliderElement.style.background = 'linear-gradient(to left, rgb(80, 80, 179), rgb(173, 216, 230), rgb(255, 185, 120), rgb(139, 90, 60))'
+      })
+    }, 10)
   }
 
   loadTargetHumidity() {
@@ -76,22 +71,26 @@ export class HumidifierDehumidifierManageComponent implements OnInit {
     this.autoHumidity = [this.targetHumidifierHumidity, this.targetDehumidifierHumidity]
   }
 
-  setTargetMode(value: number | 'off') {
+  setTargetState(value: number) {
+    this.targetState = value
+    this.service.getCharacteristic('Active').setValue(this.targetState)
+    this.loadTargetHumidity()
+  }
+
+  setTargetMode(value: number) {
     this.targetMode = value
-
-    if (this.targetMode === 'off') {
-      this.service.getCharacteristic('Active').setValue(0)
-    } else {
-      if (this.service.getCharacteristic('Active').value === 0) {
-        this.service.getCharacteristic('Active').setValue(1)
-      }
-      this.service.getCharacteristic('TargetHumidifierDehumidifierState').setValue(this.targetMode)
-    }
-
+    this.service.getCharacteristic('TargetHumidifierDehumidifierState').setValue(this.targetMode)
     this.loadTargetHumidity()
   }
 
   onHumidityStateChange() {
+    this.autoHumidity = [this.targetHumidifierHumidity, this.targetDehumidifierHumidity]
+    this.targetHumidityChanged.next(undefined)
+  }
+
+  onAutoHumidityStateChange() {
+    this.targetHumidifierHumidity = this.autoHumidity[0]
+    this.targetDehumidifierHumidity = this.autoHumidity[1]
     this.targetHumidityChanged.next(undefined)
   }
 }