+++ ui/src/app/core/accessories/types/thermostat/thermostat.manage.component.ts
@@ -1,4 +1,4 @@
-import { DecimalPipe, NgClass } from '@angular/common'
+import { DecimalPipe, NgClass, UpperCasePipe } from '@angular/common'
 import { Component, inject, Input, OnInit } from '@angular/core'
 import { FormsModule } from '@angular/forms'
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap'
@@ -9,6 +9,7 @@ import { debounceTime, distinctUntilChanged } from 'rxjs/operators'
 
 import { ServiceTypeX } from '@/app/core/accessories/accessories.interfaces'
 import { ConvertTempPipe } from '@/app/core/pipes/convert-temp.pipe'
+import { SettingsService } from '@/app/core/settings.service'
 
 @Component({
   selector: 'app-thermostat-manage',
@@ -22,13 +23,15 @@ import { ConvertTempPipe } from '@/app/core/pipes/convert-temp.pipe'
     DecimalPipe,
     TranslatePipe,
     ConvertTempPipe,
+    UpperCasePipe,
   ],
 })
 export class ThermostatManageComponent implements OnInit {
   $activeModal = inject(NgbActiveModal)
+  $settings = inject(SettingsService)
 
   @Input() public service: ServiceTypeX
-  public targetMode: any
+  public targetMode: number
   public targetTemperature: any
   public targetTemperatureChanged: Subject<string> = new Subject<string>()
 
@@ -46,6 +49,12 @@ export class ThermostatManageComponent implements OnInit {
   ngOnInit() {
     this.targetMode = this.service.values.TargetHeatingCoolingState
     this.loadTargetTemperature()
+    setTimeout(() => {
+      const sliderElements = document.querySelectorAll('.noUi-target')
+      sliderElements.forEach((sliderElement: HTMLElement) => {
+        sliderElement.style.background = 'linear-gradient(to right, rgb(80, 80, 179), rgb(173, 216, 230), rgb(255, 185, 120), rgb(139, 90, 60))'
+      })
+    }, 10)
   }
 
   loadTargetTemperature() {