+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - updates to the `manual_update_command` language string
 - show restart modal after installing an already configured plugin
 - add long click modal for lock mechanism + management
+- improvements to heater/humidifier/thermostat icon+modal
 
 ### Homebridge Dependencies